add tests for hash
